Sampled demand data + noise is added
You can choose to simulate the demand data with or without noise by setting the third argument of the environment initialization to either 0 or 1. The demand data is an average btw daily demand + noise. The noise is gaussian mean 0 with standard deviation calculated from the daily demand data for multiple realizations.