Heatmap of the last Q table
First, run the last version of the main and definitions! After that, run this code and you will get a heat map that compares the Q values for the two action in two plots. Also, you will see a plot that will show you what action will the agent choose according to that Q table.